Job Controller Put functionallity (Now the author of a job can change its properties)

+++ src/main/java/com/linkedin/service/JobService.java
@@ -2,6 +2,8 @@ package com.linkedin.service;
 
 import com.linkedin.converter.JobConverter;
 import com.linkedin.entities.database.Job;
+import com.linkedin.entities.database.Login;
+import com.linkedin.entities.database.User;
 import com.linkedin.entities.database.repo.JobRepository;
 import com.linkedin.entities.model.jobs.JobDto;
 import com.linkedin.entities.model.jobs.JobRequestDto;
@@ -12,17 +14,20 @@ import org.springframework.stereotype.Service;
 
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
 public class JobService {
     private final JobRepository jobRepository;
     private final JobConverter jobConverter;
+    private final UserService userService;
 
     @Autowired
-    public JobService(JobRepository jobRepository, JobConverter jobConverter) {
+    public JobService(JobRepository jobRepository, JobConverter jobConverter, UserService userService) {
         this.jobRepository = jobRepository;
         this.jobConverter = jobConverter;
+        this.userService = userService;
     }
 
     public Job createJob(JobRequestDto dto) {
@@ -50,15 +55,59 @@ public class JobService {
 
 
         if (existsJob(jobId)) {
+            //we check if the User that tries to erase the Job is re author of the Job
+            Login login = AuthenticationFacade.authenticatedUser();
+            Long  userId = login.getUserId();
+
+            Job job = jobRepository.findById(jobId).orElse(null);
+            if (job.getAuthorId() != userId) {
+                throw new Exception("Not Authorized to do this");
+            }
+
             jobRepository.deleteById(jobId);
         } else {
             throw new Exception("Object Not Found");
         }
 
+
     }
 
+    public JobDto updateJob(Long jobId,JobRequestDto jobRequestDto) throws Exception {
+
+    ///first we check if the job exists in our database
+        //we check if the User that tries update the Job is re author of the Job
+        if (existsJob(jobId)) {
+
+            Login login = AuthenticationFacade.authenticatedUser();
+            Long  userId = login.getUserId();
+
+            Job job = jobRepository.findById(jobId).orElse(null);
+            if (job.getAuthorId() != userId) {
+                throw new Exception("Not Authorized to do this");
+            }
+
+            //job = new Job();
+            job.setJobId(jobId);
+            job.setCompany(jobRequestDto.getCompany());
+            job.setDescription(jobRequestDto.getDescription());
+            job.setTitle(jobRequestDto.getTitle());
+
+            jobRepository.save(job);
+            return jobConverter.toJobDto(job);
+
+
+        } else {
+            throw new Exception("Object Not Found");
+        }
+
+
+    }
+
+
+
+
     public boolean existsJob(Long jobId) {
-        return jobRepository.findById(jobId) != null;
+        return jobRepository.existsById(jobId);
     }
 
     public JobDto getJob(Long jobId) throws Exception {
